magit-git: Copy Git's remote resolve behavior
This modifies magit-get-remote to the same behavior as the
corresponding function in Git.

https://github.com/git/git/blob/e4a4b31577c7419497ac30cebe30d755b97752c5/remote.c#L537

Briefly, the logic is:

If there's a branch and it has a remote, use that.
Else if the repo has exactly one remote, use that.
Else use origin.

The immediate reason for this change is so magit-push-implicitly
reflects git push's behavior.